gather/pip-requirements.txt
ipython 
ipdb
lxml
webstore-client
feedparser
